feat: Allow using files from kubeconfig
