+++ pkg/konfig/consts.go
@@ -1,0 +1,6 @@
+package konfig
+
+const (
+	filePerm = 0600
+	dirPerm  = 0700
+)
